nuke Satisfied with this implementation
Essentially, this will attempt to kill any process (on linux)
which has bound to a port greater than 6000 after the tests
have had their 'stop' calls made AND their ports are still
being bound to. The following situation came up on a few tests
on EC2, and while it may only be an EC2 issue it caused a lot
of problems:

1) `test`'s stop is called
2) stop returns successfully
3) `test` is still running and bound to 8080 (or any other port)
4) `test2` starts
5) `test2` attempts to bind to 8080
6) `test2` start fails and errors out
7) Some time later, `test` actually relinquishes port 8080

I do not know if this will resolve the EC2 problem directly,
but it is all I could think to do in an attempt.